src/main/java/kz/stanislav/voting/util/UserUtil.java
package kz.stanislav.voting.util;

import kz.stanislav.voting.persistence.model.User;
import org.springframework.security.crypto.password.PasswordEncoder;
import org.springframework.stereotype.Component;
import org.springframework.util.StringUtils;

@Component
public class UserUtil {
    public UserUtil() {
    }

    public User prepareToSave(User user, PasswordEncoder passwordEncoder) {
        String password = user.getPassword();
        user.setPassword(StringUtils.isEmpty(password) ? password : passwordEncoder.encode(password));
        user.setEmail(user.getEmail().trim().toLowerCase());

        return user;
    }
}